+++ Dotchi/Dotchi/Dotchi.csproj
@@ -165,13 +165,11 @@
     <Compile Include="Areas\Food\FoodAreaRegistration.cs" />
     <Compile Include="Areas\Food\Models\Dochi.cs" />
     <Compile Include="Controllers\AccountController.cs" />
-    <Compile Include="Controllers\FoodController.cs" />
     <Compile Include="Filters\InitializeSimpleMembershipAttribute.cs" />
     <Compile Include="Global.asax.cs">
       <DependentUpon>Global.asax</DependentUpon>
     </Compile>
     <Compile Include="Models\AccountModels.cs" />
-    <Compile Include="Models\Dochi.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
   </ItemGroup>
   <ItemGroup>
